Filter cards that aren't instant speed
